+++ packages/frontend/src/pages/admin/roles.policy-editor.vue
@@ -415,6 +415,36 @@ SPDX-License-Identifier: AGPL-3.0-only
 				</MkInput>
 			</template>
 		</XFolder>
+
+		<XFolder v-if="matchQuery([i18n.ts._role._options.canApproveEmojiRequests, 'canApproveEmojiRequests'])" v-model:policyMeta="policyMetaModel.canApproveEmojiRequests" :isBaseRole="isBaseRole" :readonly="readonly">
+			<template #label>{{ i18n.ts._role._options.canApproveEmojiRequests }}<span class="_juice">JUICE</span></template>
+			<template #valueText>{{ valuesModel.canApproveEmojiRequests ? i18n.ts.yes : i18n.ts.no }}</template>
+			<template #default="{ disabled }">
+				<MkSwitch v-model="valuesModel.canApproveEmojiRequests" :disabled="disabled">
+					<template #label>{{ i18n.ts.enable }}</template>
+				</MkSwitch>
+			</template>
+		</XFolder>
+
+		<XFolder v-if="matchQuery([i18n.ts._role._options.canApproveAvatarDecorationRequests, 'canApproveAvatarDecorationRequests'])" v-model:policyMeta="policyMetaModel.canApproveAvatarDecorationRequests" :isBaseRole="isBaseRole" :readonly="readonly">
+			<template #label>{{ i18n.ts._role._options.canApproveAvatarDecorationRequests }}<span class="_juice">JUICE</span></template>
+			<template #valueText>{{ valuesModel.canApproveAvatarDecorationRequests ? i18n.ts.yes : i18n.ts.no }}</template>
+			<template #default="{ disabled }">
+				<MkSwitch v-model="valuesModel.canApproveAvatarDecorationRequests" :disabled="disabled">
+					<template #label>{{ i18n.ts.enable }}</template>
+				</MkSwitch>
+			</template>
+		</XFolder>
+
+		<XFolder v-if="matchQuery([i18n.ts._role._options.canApproveSignups, 'canApproveSignups'])" v-model:policyMeta="policyMetaModel.canApproveSignups" :isBaseRole="isBaseRole" :readonly="readonly">
+			<template #label>{{ i18n.ts._role._options.canApproveSignups }}<span class="_juice">JUICE</span></template>
+			<template #valueText>{{ valuesModel.canApproveSignups ? i18n.ts.yes : i18n.ts.no }}</template>
+			<template #default="{ disabled }">
+				<MkSwitch v-model="valuesModel.canApproveSignups" :disabled="disabled">
+					<template #label>{{ i18n.ts.enable }}</template>
+				</MkSwitch>
+			</template>
+		</XFolder>
 	</div>
 </template>